+++ libHDiffPatch/HDiff/private_diff/match_block.h
