+++ vcdiff_wrapper/vcdiff_wrapper.cpp
@@ -27,7 +27,7 @@
  */
 #include "vcdiff_wrapper.h"
 #include "vcpatch_code_table.h"
-#include "../libHDiffPatch/HDiff/match_block.h"
+#include "../libHDiffPatch/HDiff/private_diff/match_block.h"
 #include "../libHDiffPatch/HDiff/private_diff/mem_buf.h"
 #include "../libHDiffPatch/HDiff/private_diff/limit_mem_diff/stream_serialize.h"
 #include "../libHDiffPatch/HPatch/patch.h"
